# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Gestão Escolar",
        "description": "API para gerenciamento de sistema escolar.",
        "contact": {
            "email": "suporte@exemplo.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/classroom": {
            "get": {
                "tags": [
                    "Classroom"
                ],
                "summary": "List classroom with pagination",
                "description": "Get paginated list of classrooms",
                "operationId": "472eeda86ef0afa6039764c4fd5e38a0",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term for filtering classroom",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "sort_column",
                        "in": "query",
                        "description": "Column to sort by",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "id"
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Direction for sorting (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of classroom",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ClassroomResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Classroom"
                ],
                "summary": "Create a new classroom",
                "description": "Create a new classroom",
                "operationId": "65183e8f1c2096a1f2830994fd75c836",
                "requestBody": {
                    "description": "JSON object containing new classroom data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClassroomStoreRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Classroom created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassroomResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/classroom/{id}": {
            "get": {
                "tags": [
                    "Classroom"
                ],
                "summary": "Retrieve details of a classroom by ID",
                "description": "Show details of a specific classroom",
                "operationId": "56a9ed9dbf934635736fdc41d569c683",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Classroom ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Classroom data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassroomResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Classroom"
                ],
                "summary": "Update an existing classroom",
                "description": "Update an existing classroom",
                "operationId": "ec538d0ba37774dd80aafffab975a369",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Classroom ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Updated classroom data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ClassroomUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Classroom updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ClassroomResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Classroom not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Classroom"
                ],
                "summary": "Delete a classroom by ID",
                "description": "Delete a specific classroom",
                "operationId": "aeda283cd8fa6de4b557f724bc0ffb85",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Classroom ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Classroom deleted successfully"
                    },
                    "404": {
                        "description": "Classroom not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ClassroomStoreRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "ClassroomUpdateRequest": {
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "ClassroomResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "capacity": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}